+++ CHANGELOG.md
@@ -5,4 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [0.2.0] Unreleased
+## [1.0.0] - 2021-04-06
+
+### Added
+
+- Add new helper function Test-StringIsUri
+- Add new helper function Wait-vCloudDirectorTask
+
+### Changed
+
+- Update Set-CIVMComputePolicy to use Wait-vCloudDirectorTask helper function to manage task progress.